requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.5.0
python-multipart>=0.0.9
opencv-python-headless>=4.10.0.84
Pillow>=10.0.0
imagehash>=4.3.1
easyocr>=1.7.1
numpy>=1.24.0
SQLAlchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0
ultralytics>=8.0.0
psutil>=5.9.0

